refactor: move Python code into a package
Python files are now imported as pathfinder.module
