+++ pathfinder/explore_lines.py
