+++ setup.cfg
@@ -1,12 +1,14 @@
 [semantic_release]
-version_variable=python/__init__.py:__version__
+version_variable=pathfinder/__init__.py:__version__
 changelog_components=semantic_release.changelog.changelog_table
 commit_subject=release: v{version}
 upload_to_pypi=false
-build_command=mkdir dist && tar -cvf dist/pathfinder_vim.tar.gz python/*.py plugin/*.vim LICENSE README.md
+build_command=mkdir dist && tar -cvf dist/pathfinder_vim.tar.gz pathfinder/*.py plugin/*.vim LICENSE README.md
 
 [isort]
+default_section=THIRDPARTY
 known_third_party=vim
+known_first_party=pathfinder
 order_by_type=True
 combine_star=True
 reverse_relative=True
